awyeah! Replace julia libs copying code with `-j`!
Ohcool! `juliac.jl` already has a handy switch (`-j`) to copy ("sync")
all the requisite Julia libraries into the builddir. I was doing this
manually, finding the Julia libraries directory by inspecting
`Base.julia_cmd()`.

---------

Sadly, this didn't fix my portability problem like I'd hoped... :'(
I've had this problem ever since switching to compiling my programs
instead of bundling the scripts with the entire Julia.app: When
I copy a binary compiled on this MacBook Pro over to a new mac computer,
it complains:
 > ERROR: Target architecture mismatch. Please delete or regenerate
 sys.{so,dll,dylib}.

But the crazy thing is that there *is no* sys.dylib. I already deleted
it / didn't copy it. So my guess is that somehow the compiled binary
believes itself to be the sys.dylib. But there clearly must be a way to
fix this, since lots of people compile and then distribute mac apps!