Added options to sign & sandbox finished bundle.
The cert provided to `--certificate` will be used to codesign all libs in
`<AppName>.app/Contents/MacOS/*`. Signing an app bundle with a certificate
registered with Apple is needed if you want to run the app on macOS
computers without triggering a security warning.

An entitlements file can also be specified and will be added to the app.
This is how you can "sandbox" the app, which is required to submit the
app to the Mac App Store.