Add a working Blink example app!
Okay! I've built a working app that uses Blink.jl! The program
examples/blink.jl is a working Blink app, and it runs without Julia
installed!

The one problem that it has remaining is that the OS seems to forget the
application is open, so if you double-click it again while it's already
open, the second open will hang forever and it even refused to Force
Quit (I had to restart to get it to quit). :(
I'm not sure why this is happening, but I assume it's something to do
with the fact that the application turns around and opens the embedded
Julia.app after starting, and it's then running two Mac Applications.