Fix test/examples.jl to run on Travis.
Use `mktemp()` for builddir so the tests aren't writing in the package
dir.

Change `julia build_app.jl` call to `--verbose` so that we can
see in the Travis output why it's failing.

Change build_app.jl script invocation from `@test success(cmd)` to
`@test contains(readstring(cmd), "Done building")`. This verifies that
the script finishes (the same as the first method), but on a failure it
actually prints the output, unlike `@test success` which supresses it.

--

Disable blink.jl example test because it's not yet working on Travis.