OK! Replace need for RuntimeUtils by injecting cd_to_bundle_resources() definition!
We've found a great solution to having to separate the build-time utils
and run-time utils. Instead of forcing user code to call the
change_dir_if_bundle() manually (and therefore needing to import
ApplicationBuilder in some form), we'll just /inject/ that code at
build-time, and call it from our `program.c`, always! :)

If for some reason users don't want this (but i can't see why they
wouldn't), we can always add a flag to build_app_bundle to disable it.